Refactor operations to separate files
